fix: stop reporting failed deletions as successful
DeleteFileSystem recognised the errors meaning "already gone" and let every other
error fall out of its type switch into a bare return nil. A 500, an authorization
failure, or retries exhausted against an unreachable cluster were all reported as
a successful deletion, so DeleteVolume told Kubernetes the volume was gone while
the filesystem was still on the cluster.

Its caller in volume.go was already written for this: it forgives
ObjectNotFoundError, ApiNotFoundError and ApiBadRequestError as an already-deleted
filesystem and returns Internal for anything else. It simply never received
anything else. So this restores the behaviour the caller was written against
rather than introducing new failure paths.

DeleteNfsClientGroup had the same shape and is fixed with it. It has no callers
outside its own test, so nothing changes today.

The three delete handlers now share one classifyDeleteError rather than a third
copy of the same switch. Which error codes mean "already gone" differs per object,
so they are passed in, and a test pins that a filesystem-absent code is not
forgiven for a permission.

Left alone deliberately: DeleteNfsClientGroup looks for
FilesystemDoesNotExistException, which reads like a copy from the filesystem
handler rather than anything a client group deletion returns. Changing it would
alter which errors are forgiven, and this change is only about which are reported.

Also corrects an earlier claim of mine: GetFileSystemByUid and
GetFileSystemMountToken look like they share this bug but both have a default
case returning the error, so neither swallowed anything.
